moves files into admin subdirectory
