+++ src/admin/components/app.js
