+++ src/admin/components/search_results.js
