+++ src/admin/main.js
@@ -3,8 +3,8 @@
   'use strict'
 
   var React = require('react')
-  var routes = require('./admin_routes.js')()
-  require('./styles/main.scss')
+  var routes = require('./routes.js')()
+  require('../styles/main.scss')
 
   React.render(routes, document.body)
